refactor: make datasets able to use translation hook
